Reorganize cpp_audio directory
