+++ src/cpp_audio/synths/IsochronicTone.cpp
@@ -1,5 +1,5 @@
 #include "IsochronicTone.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 #include <cmath>
 #include <algorithm>
