+++ src/cpp_audio/synths/RhythmicWaveshaping.cpp
@@ -1,5 +1,5 @@
 #include "RhythmicWaveshaping.h"
-#include "AudioUtils.h"
+#include "../utils/AudioUtils.h"
 #include <cmath>
 
 using namespace juce;
