+++ src/cpp_audio/synths/SynthFunctions.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <juce_audio_basics/juce_audio_basics.h>
-#include "Track.h"
+#include "../core/Track.h"
 
 juce::AudioBuffer<float> binauralBeat(double duration, double sampleRate, const juce::NamedValueSet& params);
 juce::AudioBuffer<float> binauralBeatTransition(double duration, double sampleRate, const juce::NamedValueSet& params);
